+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Fix schema issue in the `Event` model
+* Fix schema issue in the `Event`, `Message`, and `Draft` models
 
 ## [2.0.0] - Released 2024-02-05