+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+* Add support for getting OAuth token info
+
+### Changed
 * Fix schema issue in the `Event`, `Message`, and `Draft` models
 
 ## [2.0.0] - Released 2024-02-05