+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Add support for getting OAuth token info
 
 ### Changed
-* Fix schema issue in the `Event`, `Message`, and `Draft` models
+* Fix schema issues in the `Event`, `Message`, `Draft`, and `CodeExchangeResponse` models
 
 ## [2.0.0] - Released 2024-02-05