+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Nylas Java SDK Changelog
 
-## Unreleased
+## [2.1.0] - Released 2024-02-12
 
 ### Added
-
 * Add support for getting OAuth token info
 
 ### Changed
@@ -412,7 +411,8 @@ This second release aims toward API stability so that we can get to v1.0.0.
 
 Initial preview release
 
-[Unreleased]: https://github.com/nylas/nylas-java/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/nylas/nylas-java/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/nylas/nylas-java/releases/tag/v2.1.0
 [2.0.0]: https://github.com/nylas/nylas-java/releases/tag/v2.0.0
 [1.21.0]: https://github.com/nylas/nylas-java/releases/tag/v1.21.0
 [1.20.1]: https://github.com/nylas/nylas-java/releases/tag/v1.20.1